requirements.txt
click
prometheus_client
pyyaml
